+++ ElvUI_MerathilisUI/Core/Changelog/6.48.lua
@@ -1,0 +1,10 @@
+local MER = unpack(ElvUI_MerathilisUI)
+
+MER.Changelog[647] = {
+	RELEASE_DATE = "19.12.2024",
+	FIXES = {
+		"[Group Info]: C_LFGList.GetActivityInfoTable got changed with 11.0.7"
+	},
+	NEW = {},
+	IMPROVEMENTS = {},
+}