+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ### Changes ###
 
-  * [Misc]: Bump toc
+  * [Group Info]: C_LFGList.GetActivityInfoTable got changed with 11.0.7